Added a recursive definition for executing a small step.
It returns TPHang if no small step can be executed.  This is currently
not consistent with the inductive definition, which does not implement
any handling of hanging excutions.

The best way to fix this might be create a definition that clearly
states, that a small step is only possible, if the result of applying
smallstep to the expression is *not* TPHang.  Using this definition, it
should be possible to show, that the inductive definition is consistent
with the recursive definition of small steps.