A lot of changes: - split the project into more files. I don't know if this makes things   better or worse, due to Coq's insane way of handling namespaces and   imports this will probably be very hard to maintain and extend. If any   on you have a better idea, feel free to change it. - replaced the TPSubst function with the TPSubstitutesInto predicate - introduced predicates TPIsFreeIn and TPIsNotFreeIn - implemented proper alpha conversion for smallsteps (via axiomatic   rename function) - replaced implementation of typing environments as lists with abstract   environments that satisfy some reasonable axioms - proved various lemmas concerning the interplay between typing,   substitution, smallsteps and environment manipulation - proved the preservation theorem - started on the progress theorem
The definition of smallsteps, substitution and free variables should be
okay now (I was able to prove preservation after all ;-) ). Nevertheless
there are still MANY things to work on (i.e. a lot of 'admit's).

I'll try to whip up some documentation in the next days. I finished the
proof of preservation just last night and didn't want to wait any
longer, so the code is still in a state of disarray I'm afraid.